call syscalls with the nix library
Instead of using the portable functions that are part of
the standard library, use "nix" which provides safe access to pipe,
fork, exec, and other posix syscalls that are necessary for writing
efficient shells.

The std::process module doesn't provide any way to
make pipes in a cross-platform way. Doing the equivalent of forking
on windows is also complicated.